+++ src/objects/user.py
@@ -22,7 +22,7 @@ from datetime import datetime as dt
 
 import redis
 
-from OrarioTreniBot import config
+from .. import config
 
 r = redis.StrictRedis(host=config.REDIS_HOST, port=config.REDIS_PORT, db=config.REDIS_DB)
