+++ start.py
@@ -18,7 +18,7 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-from OrarioTreniBot import main
+from src import main
 
 if __name__ == "__main__":
     main.bot.run()
